feat: Add API to test Redis cluster data sharding
The code changes introduce a new API endpoint `/test-availability` in the `TestController` class. This API is used to test the availability and automatic failover of Redis Cluster. It verifies if the system remains functional when a master node fails. The API performs the following steps:
- Writes data to the cluster
- Manually shuts down one of the master nodes
- Attempts to read data from the cluster

This commit adds the necessary code to implement this functionality.